#308 Remove unnecessary casts

+++ src/ca/mcgill/cs/jetuml/views/edges/SegmentationStyleFactory.java
@@ -344,11 +344,11 @@ public final class SegmentationStyleFactory
 						
 			if( pSide.isEastWest() )
 			{		
-				return (int)(otherNode1.view().getBounds().getCenter().getY() - otherNode2.view().getBounds().getCenter().getY());
+				return otherNode1.view().getBounds().getCenter().getY() - otherNode2.view().getBounds().getCenter().getY();
 			}
 			else
 			{
-				return (int)(otherNode1.view().getBounds().getCenter().getX() - otherNode2.view().getBounds().getCenter().getX());
+				return otherNode1.view().getBounds().getCenter().getX() - otherNode2.view().getBounds().getCenter().getX();
 			}
 		});
 	}
